Removed the client loop and added a basic send
Removed the client loop as it isnt really useful at the moment untill
feedback from the server is implmented and is not the intended method to
communicate with the server.

It was replaced with a sendmessage method that just send a single
message to the server. The main method now reflect this allowing a user
to send a single agrument (can be quoted) to the server directly.