Inversed the check for data
changed the check to see if any data was sent in the connection to check
for the lack of data (not data) and to print a error message and
continue the loop.

This is more flexable then placing everything inside the positive check
making it easier to follow the code.