Moved the exit code to process input
Removed the quit break in the daemon loop and added a exit to the
process input method for quit. This removes spical case code from the daemon
loop and keeps it in the process_input method.

Also changed the process_input to stop the mc server as it no longer has
a quit method.