Allowed the download to run while mc is
Made it so that the check to see if the minecraft server is running when
upgrading to be moved to the end and only change the return message.
This allows users to upgrade the minecraft server while it is running
but tells them to restart when it is done.

Overall this should lower the downtime of minecraft servers as the user
can upgrade then restart the server rather then stopping, waitin for the
downlaod and starting it again.