Added a reply from the server to client
Added a method to reply from the server to the client so that the client
has better information about what is happening to the server. All the
process_input cases now use this method to communicate with the client
to inform them of whats happening.

The protocal used to communicate is as follows:
client sends a message to the server,
server replys with a message to be displayed to the user or a message to
be interperted by the client.
server keeps replying until it sends a termination message.

Messages to be interperated by the client should be full caps, currently
implmented messages are:
SET TIMEOUT X  - set the timeout of the client to X (used if the server
knows that it isnt going to repond for a while)
DONE - Tells the client the server will no longer send any messages and
allows the client to exit.