Added the networking module
This module is ment to replace server.py supplying the server and client
programs. It uses socketserver rather then just the raw socket classes
to provide and easier and more maintainable solution.

The server and client are also cleanly split into two classes (Server
and Client) respectivly making it easier to see which is which.