Fixed error: sending a string starting with stop
Fixed an error when sending a string that starts with stop but isnt just
'stop' causing the minecraft server to stop, but not pycraft to tidy
up after it exits (by nullifing _mcp). This caused pycraft to think the
minecraft server was still running even when it wasne't thus stopping
being able to start it again.